+++ arrow-fx-test/build.gradle
@@ -14,8 +14,6 @@ apply from: "$PUBLISH_CONF"
 dependencies {
     compile project(":arrow-fx")
     api "io.arrow-kt:arrow-core-test:$VERSION_NAME", excludeArrow
-    compile "io.arrow-kt:arrow-mtl:$VERSION_NAME", excludeArrow
-    compile "io.arrow-kt:arrow-free:$VERSION_NAME", excludeArrow
     compile "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$KOTLIN_VERSION"
     compile "org.jetbrains.kotlinx:kotlinx-coroutines-core:$KOTLINX_COROUTINES_VERSION"
     testRuntime "org.junit.vintage:junit-vintage-engine:$JUNIT_VINTAGE_VERSION"
