+++ arrow-fx-test/src/main/kotlin/arrow/test/laws/Law.kt
@@ -11,11 +11,6 @@ import io.kotlintest.properties.Gen
 import io.kotlintest.should
 import io.kotlintest.shouldNot
 
-// TODO: check to move to arrow-core-test
-fun throwableEq() = Eq { a: Throwable, b ->
-  a::class == b::class && a.message == b.message
-}
-
 fun <A> A.equalUnderTheLaw(b: A, eq: Eq<A>): Boolean =
   shouldBeEq(b, eq).let { true }
