+++ arrow-fx/src/main/kotlin/arrow/fx/typeclasses/Fiber.kt
@@ -11,7 +11,8 @@ import arrow.higherkind
  * You can think of fibers as being lightweight threads, a Fiber being a
  * concurrency primitive for doing cooperative multi-tasking.
  */
-@higherkind interface Fiber<F, out A> : FiberOf<F, A> {
+@higherkind
+interface Fiber<F, out A> : FiberOf<F, A> {
 
   /**
    * Returns a new task that will await for the completion of the
